alex/bindings: sticker shape
