--- apps/examples/src/examples/pin-example/README.md
@@ -1,9 +1,0 @@
----
-title: Pin (bindings)
-component: ./PinExample.tsx
-category: shapes/tools
----
-
-Pin shapes together using bindings
-
----
