+++ apps/examples/src/examples/sticker-bindings/README.md
@@ -1,0 +1,9 @@
+---
+title: Sticker (bindings)
+component: ./StickerExample.tsx
+category: shapes/tools
+---
+
+A sticker shape, using bindings to attach shapes to one and other
+
+---
